create buy_sell_hold function that creates label for marchine learning algorithm
The idea is to map this function to a pandas dataframe column and that column will be the label
pass in any number of columns (*args). in practice these columns will be our future price change columns that were processed in process_data_for_labels.

If the price went up 2% it is a buy, if it went down 2% it is a sell, else it is a hold